Reorganized extension code to cask and Play JSON so that it is centralized under the util.helpers package.
