+++ family-recipes/src/main/scala/api/package.scala
@@ -1,9 +1,7 @@
-import io.undertow.util.HeaderMap
-import play.api.libs.json.{ Format, JsValue, Json, Reads }
-import util._
+import play.api.libs.json.{ Format, JsValue, Json }
+import util.helpers.jsonExt._
 
 import scala.language.implicitConversions
-import scala.util.Try
 
 package object api {
 
@@ -16,17 +14,6 @@ package object api {
   ) extends Exception(statusText.orNull)
 
   val AuthenticationError: ApiError = ApiError(403)
-
-  implicit class RequestExt(request: cask.Request) {
-    def as[T: Reads]: T =
-      Try(Json.parse(request.text()).as[T]).getOrElse {
-        throw api.ApiError(400, Some(s"Malformed JSON"), Some(request.text()))
-      }
-  }
-
-  implicit class HeaderValuesExt(headerMap: HeaderMap) {
-    def getHeader(fieldName: String): Option[String] = Option(headerMap.get(fieldName)).flatMap(header => Option(header.getFirst))
-  }
 
   case class Response(statusCode: Int, statusText: Option[String] = None, data: Option[JsValue] = None)
