+++ family-recipes/src/main/scala/util/package.scala
@@ -1,9 +1,6 @@
 import net.ceedubs.ficus.Ficus._
 import org.apache.commons.codec.binary.Base64
 import org.apache.commons.codec.digest._
-import play.api.libs.json.{ JsValue, Json, Writes }
-
-import scala.language.implicitConversions
 
 package object util {
 
@@ -14,13 +11,4 @@ package object util {
     Base64.encodeBase64URLSafeString(hmac)
   }
 
-  implicit class JsonExt[T](data: T)(implicit writes: Writes[T]) {
-    def asJson: JsValue = Json.toJson(data)
-  }
-
-  implicit class JsValueExt(json: JsValue) {
-    def asString: String = json.toString()
-  }
-
-  implicit def asString(json: JsValue): String = json.toString
 }
